\encoding{latin1} 
\name{multiColl-package}
\alias{multiColl-package}
\alias{multiColl}
\docType{package}
\title{
  Collinearity detection in a multiple linear regression model.
}
\description{
  R package to detect collinearity in a multiple linear regression model.
}
\details{
  The detection of worrying approximate multicollinearity in a multiple linear regression model is a problem addressed in all existing statistical packages. However, we have detected deficits regarding to the incorrect treatment of qualitative independent variables and the role of the intercept of the model. The objective of this package is to correct these deficits. In this package will be available detection and treatment techniques traditionally used as the recently developed.
}
\author{
  Román Salmerón Gómez (University of Granada), Catalina García García (University of Granada) and José García García (University of Almería).
}
\references{
  \code{multiColl}: An R Package for Detecting Multicollinearity. Working paper.
}
 
